api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
snowflake-connector-python==3.12.2
python-dotenv==1.0.1
confluent-kafka==2.5.3
pandas==2.2.2
